refactor: renomme le module
