+++ test/test_settle.py
@@ -1,8 +1,8 @@
 import random
 
-from bookkeeper.balance import balance
-from bookkeeper.definitions import Member, Transaction
-from bookkeeper.settle import settle
+from expenses.balance import balance
+from expenses.definitions import Member, Transaction
+from expenses.settle import settle
 
 
 def test_simple_settlement():
